server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
scikit-learn>=1.3.0
numpy>=1.24.0
sentence-transformers>=2.2.0
feedparser>=6.0.0
requests>=2.31.0
groq>=0.4.0
python-dotenv>=1.0.0
